refactor: split window.rs into a module
